learnopengl section 1 finished! + gamewindow, inputmanager
input manager is a mixture of callbacks and data polled at the beginning of the frame, which all culminates in a very easy to use, extensible input system. the only bug i've found so far is that, if you press a button EXTREMELY fast, it'll only recognize a release input instead of a press input.
gamewindow is a basic context, and every time it's initialized, it creates a new window, and assigns a pointer to itself as the void pointer on the glfw window. it then has a static helper function to cast the void pointer as a game window! for now, it's fine as is, as i don't plan on having multiple windows (or at least not multiple windows that aren't abstracted using the gamewindow class), so i don't need any elaborate checks to see if the void pointer is valid or not :]