repurpose Scene, refactor to separate functionality, remove boost
hoo boy! i've been doing a TON of reading on ecs and i'm starting to finally get it. it's so seamless, and the code already looks so pretty and bug free :]]

and apparently boost increases compile times more noticeably on smaller projects, and i don't use it anymore so c'est la vie